pyproject.toml
[project]
name = "mapvoyage-extract"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiobotocore>=2.24.2",
    "aiofiles>=24.1.0",
    "aiohttp>=3.11.16",
    "asyncssh>=2.20.0",
    "mwparserfromhell>=0.6.6",
    "python-dotenv>=1.1.0",
    "wikitextparser>=0.56.3",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]
